requirements.txt
langchain
langchain-community
chromadb
gradio
python-dotenv
pandas
sentence-transformers
langchain-google-genai
